refac: Refactor app_data.rs into modular components
Split the monolithic app_data.rs file into smaller, focused modules:
- container.rs: Container status and state
- service.rs: Service port mapping
- ttl.rs: Application time-to-live handling
- settings.rs: Application settings
- status.rs: Application status management
- data.rs: Main AppData implementation
